refactor: migrate useSidebar into useLayout and useSidebarControl
