+++ src/client/theme-default/composables/aside.ts
@@ -1,9 +1,9 @@
 import { useMediaQuery } from '@vueuse/core'
 import { computed } from 'vue'
-import { useSidebar } from './sidebar'
+import { useLayout } from './layout'
 
 export function useAside() {
-  const { hasSidebar } = useSidebar()
+  const { hasSidebar } = useLayout()
   const is960 = useMediaQuery('(min-width: 960px)')
   const is1280 = useMediaQuery('(min-width: 1280px)')
